requirements.txt
matplotlib==3.3.1
nltk==3.8.1
numpy==1.26.3
pandas==2.2.3
Pillow==11.0.0
python-dotenv==1.0.1
Requests==2.32.3
wordcloud==1.9.3
